refactor(test): move DatabaseCleanup into test package as it is only for test
