+++ src/test/java/nextstep/subway/acceptance/DatabaseCleanup.java
@@ -1,4 +1,4 @@
-package nextstep.subway.application;
+package nextstep.subway.acceptance;
 
 import com.google.common.base.CaseFormat;
 import java.util.List;
